added monthly listings and total listing and best sellers email address to final report class

+++ src/main/java/com/worldofbooks/exercise/model/Report.java
@@ -14,7 +14,7 @@ import java.util.List;
 @AllArgsConstructor
 public class Report {
 
-    private Integer totalListing;
+    private Long totalListing;
 
     private Integer totalEbayListingNumber;
